Finance item fragment

+++ app/src/main/java/com/example/coinkeeper/domain/FinanceItem.kt
@@ -3,7 +3,7 @@ package com.example.coinkeeper.domain
 data class FinanceItem(
     val name: String,
     val comment: String,
-    val sum: Double,
+    val sum: Int,
     val category: Int,
     var id:Int = UNDEFINED_ID
 ) {
